+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to the "quotify" extension will be documented in this file.
 
 ## [0.0.3]
-- Added simple command to convert a group of lines to a JS style array / SQL style set
+- Added simple command to convert a group of lines to a JS style array and comma newline delmited text
+- Have a sql list but need more robust testing around it and escaping single ticks
 
 ## [0.0.2]
 - Changes to package only no code changes